Fix latest problem

+++ FinanceManager/Services/StorageServices.cs
@@ -1,36 +1,45 @@
-﻿using FinanceManager.Models;
+﻿using FinanceManager.DTO;
+using FinanceManager.Models;
 using Microsoft.EntityFrameworkCore;
 
 namespace FinanceManager.Services
 {
     public class StorageServices
     {
-        private readonly Context _context = new Context();
+        private readonly Context _context;
 
-        public async Task<Storage?> GetStorageAsync(int? id)
+        private StorageServices()
         {
-            if (id is null)
-            {
-                return null;
-            }
 
-            return await _context.Storages
-                .Include(t=> t.Transactions)
-                .FirstOrDefaultAsync(x => x.Id == id);
         }
 
-        public async Task<List<Storage>> GetAllStoragesAsync()
+        public StorageServices(Context context) => _context = context;
+
+        public async Task<List<Storage>> GetAllAsync()
         {
             return await _context.Storages
                 .ToListAsync();
         }
 
-        public async Task<bool> CreateStorageAsync(Storage storage)
+        public async Task<Storage?> GetAsync(int id)
         {
-            if(!StorageIsValid(storage))
+            return await _context.Storages
+                .Include(t=> t.Transactions)
+                .FirstOrDefaultAsync(x => x.Id == id);
+        }
+
+        public async Task<bool> CreateAsync(StorageCreateDTO storageData)
+        {
+            if(!IsValid(storageData))
             {
                 return false;
             }
+
+            var storage = new Storage()
+            {
+                Name = storageData.Name,
+                Value = storageData.Value
+            };
 
             await _context.Storages.AddAsync(storage);
             await _context.SaveChangesAsync();
@@ -38,9 +47,9 @@ namespace FinanceManager.Services
             return true;
         }
 
-        public async Task<bool> EditStorageAsync(int? id, Storage storage)
+        public async Task<bool> EditAsync(int id, StorageUpdateDTO storageData)
         {
-            if(id!= storage.Id) 
+            if(id!= storageData.Id) 
             {
                 return false;
             }
@@ -48,6 +57,17 @@ namespace FinanceManager.Services
             if(!_context.Storages.Any(x=> x.Id  == id)) 
             {
                 return false;
+            }
+
+            var storage = await _context.Storages.FirstAsync(x=> x.Id == id);
+
+            if(!string.IsNullOrEmpty(storageData.Name) & storage.Name != storageData.Name)
+            {
+                storage.Name = storageData.Name;
+            }
+            if(storage.Value != storageData.Value)
+            {
+                storage.Value = storageData.Value;
             }
 
             try
@@ -63,13 +83,8 @@ namespace FinanceManager.Services
             return true;
         }
 
-        public async Task<bool> DeleteStorageAsync(int? id)
+        public async Task<bool> DeleteAsync(int id)
         {
-            if(id is null)
-            {
-                return false;
-            }
-
             var storage = await _context.Storages
                 .Include(t => t.Transactions)
                 .FirstOrDefaultAsync(s => s.Id == id);
@@ -79,17 +94,23 @@ namespace FinanceManager.Services
                 return false;
             }
 
+            if(storage.Transactions.Count >0)
+            {
+                throw new Exception("This storage contain transactions");
+            }
+
             _context.Storages.Remove(storage);
             await _context.SaveChangesAsync();
             return true;
         }
 
-        private bool StorageIsValid(Storage storage)
+        private bool IsValid(StorageCreateDTO storageData)
         {
-            if(_context.Storages.FirstOrDefault(s=> s.Name == storage.Name) != null & storage.Name == null)
+            if (storageData.Name == null & _context.Storages.FirstOrDefault(s=> s.Name == storageData.Name) != null)
             {
                 return false;
             }
+
             return true;
         }
     }
